+++ src/main/kotlin/com/example/mykku/dailymessage/DailyMessageService.kt
@@ -1,4 +1,4 @@
-package com.example.mykku.dailymessage.service
+package com.example.mykku.dailymessage
 
 import com.example.mykku.dailymessage.domain.DailyMessage
 import com.example.mykku.dailymessage.domain.SortDirection
