+++ src/main/kotlin/com/example/mykku/feed/FeedCommentService.kt
@@ -1,6 +1,9 @@
-package com.example.mykku.feed.service
+package com.example.mykku.feed
 
-import com.example.mykku.feed.dto.*
+import com.example.mykku.feed.dto.CommentAuthorResponse
+import com.example.mykku.feed.dto.FeedCommentReplyResponse
+import com.example.mykku.feed.dto.FeedCommentResponse
+import com.example.mykku.feed.dto.FeedCommentsResponse
 import com.example.mykku.feed.tool.FeedCommentReader
 import com.example.mykku.feed.tool.FeedReader
 import com.example.mykku.like.tool.LikeFeedCommentReader
@@ -18,10 +21,10 @@ class FeedCommentService(
     fun getComments(feedId: Long, memberId: String?, pageable: Pageable): FeedCommentsResponse {
         val feed = feedReader.getFeedById(feedId)
         val commentsPage = feedCommentReader.getCommentsByFeed(feed, pageable)
-        
+
         // 모든 부모 댓글의 대댓글을 한 번에 조회
         val repliesMap = feedCommentReader.getRepliesByParentComments(commentsPage.content)
-        
+
         val commentResponses = commentsPage.content.map { comment ->
             val replies = repliesMap[comment.id] ?: emptyList()
             val replyResponses = replies.map { reply ->
@@ -39,7 +42,7 @@ class FeedCommentService(
                     updatedAt = reply.updatedAt
                 )
             }
-            
+
             FeedCommentResponse(
                 id = comment.id!!,
                 content = comment.content,
@@ -56,7 +59,7 @@ class FeedCommentService(
                 updatedAt = comment.updatedAt
             )
         }
-        
+
         return FeedCommentsResponse(
             comments = commentResponses,
             totalElements = commentsPage.totalElements,
