+++ src/main/kotlin/com/example/mykku/feed/FeedController.kt
@@ -1,12 +1,10 @@
-package com.example.mykku.feed.controller
+package com.example.mykku.feed
 
 import com.example.mykku.auth.config.CurrentMember
 import com.example.mykku.common.dto.ApiResponse
 import com.example.mykku.exception.ErrorCode
 import com.example.mykku.exception.MykkuException
 import com.example.mykku.feed.dto.*
-import com.example.mykku.feed.service.FeedCommentService
-import com.example.mykku.feed.service.FeedService
 import com.example.mykku.member.domain.Member
 import jakarta.validation.Valid
 import org.springframework.data.domain.PageRequest
@@ -29,12 +27,12 @@ class FeedController(
         @CurrentMember member: Member
     ): ResponseEntity<ApiResponse<CreateFeedResponse>> {
         val imageList = images ?: emptyList()
-        
+
         // 이미지 개수 제한 검증
         if (imageList.size > CreateFeedRequest.MAX_IMAGE_COUNT) {
             throw MykkuException(ErrorCode.FEED_IMAGE_LIMIT_EXCEEDED)
         }
-        
+
         val request = CreateFeedRequest(
             title = requestDto.title,
             content = requestDto.content,
