+++ src/main/kotlin/com/example/mykku/like/LikeService.kt
@@ -1,4 +1,4 @@
-package com.example.mykku.like.service
+package com.example.mykku.like
 
 import com.example.mykku.board.tool.BoardReader
 import com.example.mykku.dailymessage.tool.DailyMessageCommentReader
